python-selenium-runner-example/pyproject.toml
[project]
name = "python-selenium-runner-example"
version = "0.1.0"
description = "BrikByteOS Selenium Grid runner example (pytest + RemoteWebDriver)"
requires-python = ">=3.10"
dependencies = [
  "pytest>=8.2.0",
  "selenium>=4.27.0",
  "python-dotenv>=1.0.1",
  "requests>=2.32.0",
]

[tool.pytest.ini_options]
testpaths = ["src"]
addopts = "-q"
